LPD-61435 Introduce base class to remove duplication
